rename trashcan to parentTrashcan
